+++ src/main/kotlin/us/frollo/frollosdk/authentication/AuthenticationType.kt
@@ -40,12 +40,14 @@ sealed class AuthenticationType {
      * @param authorizationUrl URL of the OAuth2 authorization endpoint for web based login
      * @param tokenUrl URL of the OAuth2 token endpoint for getting tokens and native login
      * @param revokeTokenURL OAuth2 token endpoint to revoke refresh token on logout (if supported)
+     * @param audienceUrl OAuth2 audience URL (Optional)
      */
     class OAuth2(
         val redirectUrl: String,
         val authorizationUrl: String,
         val tokenUrl: String,
-        val revokeTokenURL: String? = null
+        val revokeTokenURL: String? = null,
+        val audienceUrl: String? = null
     ) : AuthenticationType() {
 
         /**